+++ src/helpers/SubjectType.js
@@ -1,9 +1,0 @@
-export default class SubjectType {
-    static private() {
-        return 'private';
-    }
-
-    static public() {
-        return 'public';
-    }
-}