+++ src/index.js
@@ -1,6 +1,0 @@
-import Rxmq from 'rxmq';
-import applyMiddleware from 'rxmq.middleware';
-// inject middleware
-applyMiddleware();
-
-export default Rxmq;